route activated targets before sacrifice costs
